* server/character.h, server/impl.cpp, server/movie.h,                            server/movie_def_impl.cpp, server/movie_def_impl.h,           server/movie_definition.h, server/movie_interface.h,           server/movie_root.cpp, server/movie_root.h,                                     server/sprite_definition.cpp, server/sprite_definition.h,           server/sprite_instance.cpp, server/sprite_instance.h,           server/swf/ASHandlers.cpp:           Implemented incremental loading of movie (streaming);           fixed lots of references to frame numbers to be expressed           with size_t type rather then int.

+++ server/character.h
@@ -216,7 +216,7 @@ public:
 //	printf("%s:\n", __PRETTY_FUNCTION__); // FIXME:
     }	// for buttons and sprites
 
-    virtual void	goto_frame(int /*target_frame*/) {}
+    virtual void	goto_frame(size_t /*target_frame*/) {}
 
     virtual bool	get_accept_anim_moves() const { return true; }
